modified for namespace vt and use Vt gem
